Stop accepting an argument for the `--auto` switch
Being able to disable auto-switch whilst enabling dark-mode is confusing
and unimplemented, yet the possibility is implied by being able to call:

    ./dark-mode.scpt --auto=false true

The ability to do this would be crappy program design, not to mention an
incredible pain-in-the-arse to implement. Better if we operate in either
the context of auto-switching or the dark-mode setting proper.

(Whose idea was it to make these separate preferences?)